refactor: move Consumer, make setCredential its member fn
